renames internal/autobind package
renamed to internal/pgadmin based on review feedback
