+++ internal/pgadmin/doc.go
@@ -1,7 +1,7 @@
-/* package autobind provides a set of tools for interacting with the sqlite
+/* package pgadmin provides a set of tools for interacting with the sqlite
 database which powers pgadmin */
 
-package autobind
+package pgadmin
 
 /*
  Copyright 2017 - 2020 Crunchy Data Solutions, Inc.
